+++ src/SettingsView.vue
@@ -175,7 +175,7 @@ export default {
 
 
     get_group_config(config) {
-      let keys = ['on_idle', 'cpus', 'gpus', 'beta', 'key', 'cuda']
+      let keys = ['on_idle', 'cpus', 'gpus', 'beta', 'key', 'cuda', 'hip']
       let copy = copy_keys(config, keys)
 
       copy.on_idle = !!copy.on_idle
@@ -183,6 +183,7 @@ export default {
       copy.beta    = !!copy.beta
       copy.key     = copy.key || 0
       copy.cuda    = copy.cuda == undefined ? true : copy.cuda
+      copy.hip     = copy.hip == undefined ? true : copy.hip
 
       if (this.$util.version_less('8.3.1', this.version)) {
         copy.on_battery = !!config.on_battery